move steps.js into lib/
